 \name{Open_SSD}
 \alias{Open_SSD}
 \title{Open SNP set data file (SSD) }
 \description{
 Open_SSD opens the SNP Set data file (SSD).
 After finishing using the SSD file, you must close it by callinsg Close_SSD function. 
 }
 \usage{
	Open_SSD(File.SSD, File.Info)
 }
\arguments{
      \item{File.SSD}{ the name of the SSD file . }
      \item{File.Info}{ the name of the SSD info file. }
}
\value{
 	Open_SSD returns a list object of SSD.INFO which has set information. 
}


\author{Seunggeun Lee, Larisa Miropolsky}

