\name{SKAT.example.ChrX}
\alias{SKAT.example.ChrX}
\title{Example data for SKAT}
\description{
Example data for SKAT.
}
\format{
  SKAT.example contains the following objects:
  \describe{
    \item{Z}{a numeric genotype matrix of 2000 individuals and 52 SNPs. 
	Each row represents a different individual, and each column represents a different SNP marker. }
    \item{x1}{a numeric vector of continuous covariates. }
    \item{x2}{a numeric vector of binary covariates. }
    \item{Gender}{a numeric vector of gender (male=1, female=2). }
    \item{y}{a numeric vector of binary phenotypes.}
    \item{Z.A}{a list object of 10 genotype matrices.}
  }
}

