\name{SKAT.example}
\alias{SKAT.example}
\title{Example data for SKAT}
\description{
Example data for SKAT.
}
\format{
  SKAT.example contains the following objects:
  \describe{
    \item{Z}{a numeric genotype matrix of 2000 individuals and 67 SNPs. 
	Each row represents a different individual, and each column represents a different SNP marker. }
    \item{X}{a numeric matrix of 2 covariates. }
    \item{y.c}{a numeric vector of continuous phenotypes.}
    \item{y.b}{a numeric vector of binary phenotypes.}
  }
}

