\name{SKAT.fam.example}
\alias{SKAT.fam.example}
\title{Example data for SKAT_NULL_emmaX}
\description{
Example data for SKAT_emmaX.
}
\format{
  Example contains the following objects:
  \describe{
    \item{Z}{a numeric genotype matrix 
	Each row represents a different individual, and each column represents a different SNP marker. }
    \item{X}{a numeric matrix of covariates. }
    \item{y}{a numeric vector of continuous phenotypes.}
    \item{K}{a kinship matrix.}
  }
}

