\name{SSD_FILE_OPEN}
\alias{SSD_FILE_OPEN.isOpen}
\alias{SSD_FILE_OPEN.FileName}
\alias{Get_Lambda}
\alias{Get_Liu_PVal}
\alias{SKAT_Check_Method}
\alias{SKAT_Optimal_Each_Q}
\alias{SKAT_Optimal_PValue_Davies}
\alias{SKAT_Optimal_PValue_Liu}
\alias{Get_Davies_PVal}
\title{Internal variables for SSD and functions for other packages}
\description{
Internal variable for SSD and functions for other packages
}

