 \name{Open_SSD}
 \alias{Open_SSD}
 \title{Open SNP set data file (SSD) }
 \description{
 Open a SNP Set data file (SSD). After finishing using the SSD file, you must close the file by callinsg Close_SSD function. 
 }
 \usage{
	Open_SSD(File.SSD, File.Info)
 }
\arguments{
      \item{File.SSD}{ name of the SSD file . }
      \item{File.Info}{ name of the SSD info file. }
}
\value{
 	 a list object of SSD.INFO.
}


\author{Seunggeun Lee, Larisa Miropolsky}

